% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cornea.mask.R
\name{cornea.mask}
\alias{cornea.mask}
\title{cornea.mask}
\usage{
cornea.mask(obj)
}
\arguments{
\item{obj}{CanalogramImages object of images to rescale}
}
\value{
CanalogramImages object of rescaled images

Remove all (high-resolution) data from the region of the cornea using the
cornea mask.
}
\description{
cornea.mask
}
\examples{
\dontrun{
obj <- cornea.mask(images)
}
}

