% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delta.gam.R
\name{delta.gam}
\alias{delta.gam}
\title{delta.gam}
\usage{
delta.gam(before, after)
}
\arguments{
\item{before}{CanaogramGAM object before intervention}

\item{after}{CanaogramGAM object after intervention}
}
\value{
DeltaCanaogramGAM object describing the change in flow
}
\description{
delta.gam
}
\examples{
\dontrun{
images.before <- read.images('./data/before_T00', n = 5)
images.after  <- read.images('./data/after_T00', n = 5)
fit.before <- fit.gam(images.before)
fit.after <- fit.gam(images.after)
delta.fit <- get.change(fit.before, fit.after)
}
}

