% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.CanalogramImages.R
\name{plot.CanalogramImages}
\alias{plot.CanalogramImages}
\title{plot.CanalogramImages}
\usage{
\method{plot}{CanalogramImages}(x, frames = c(1,
  round(EBImage::numberOfFrames(x$data.low)/2),
  max(EBImage::numberOfFrames(x$data.low))), ...)
}
\arguments{
\item{x}{An CanalogramImages object}

\item{frames}{The image frames to display}

\item{...}{<not used>}
}
\value{
A ggplot2 object of the plots
}
\description{
plot.CanalogramImages
}
\examples{
\dontrun{
images <- read.images('./data/Trial-1/one gree_T00', n = 10)
plot(images, frames = c(1, 10))
plot(images)
}
}

