% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.CanaogramGAM.R
\name{plot.CanaogramGAM}
\alias{plot.CanaogramGAM}
\title{plot.CanaogramGAM}
\usage{
\method{plot}{CanaogramGAM}(x, ...)
}
\arguments{
\item{x}{A CanaogramGAM object}

\item{...}{<not used>}
}
\value{
A ggplot2 object with the plot
}
\description{
plot.CanaogramGAM
}
\examples{
\dontrun{
images <- read.images('./data/Trial-1/one gree_T00', n = 10)
fit <- fit.gam(images)
plot(fit)
}
}

