% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delta.gam.R
\name{plot.DeltaCanaogramGAM}
\alias{plot.DeltaCanaogramGAM}
\title{plot.DeltaCanalogramGAM}
\usage{
\method{plot}{DeltaCanaogramGAM}(x, which = c("rate", "time", "intensity",
  "all"), max.time = 10, max.rate = 5, ...)
}
\description{
plot.DeltaCanalogramGAM
}

