% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.grid.R
\name{plot.grid}
\alias{plot.grid}
\title{plot.grid}
\usage{
\method{plot}{grid}(x, which = c("rate", "time", "intensity", "all"), ...)
}
\arguments{
\item{x}{CanaogramGAM to plot}

\item{which}{Which of 'rate', 'time', 'intensity', or 'all' to plot}

\item{...}{<not used>}
}
\value{
A ggplot2 object of the plot
}
\description{
plot.grid
}
\examples{
\dontrun{
images <- read.images('./data/Trial-1/one gree_T00', n = 10)
fit <- fit.gam(images)
plot.grid(fit)
}
}

