% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.images.R
\name{plot.images}
\alias{plot.images}
\title{plot.images}
\usage{
\method{plot}{images}(x, frames = c(1, round(max(x$gam$data$t)/2),
  max(x$gam$data$t)), scale = c("scaled", "raw"), masked = FALSE,
  cornea = FALSE, ...)
}
\arguments{
\item{x}{CanaogramGAM to plot}

\item{frames}{Which frames to show}

\item{scale}{Which image scale to use (raw intensities or scalled intensities)}

\item{masked}{Show just the masked data or all data}

\item{cornea}{Include the cornea in the output (or omit it)}

\item{...}{<not used>}
}
\value{
A ggplot2 object of the plot
}
\description{
plot.images
}
\examples{
\dontrun{
images <- read.images('./data/Trial-1/one gree_T00', n = 10)
fit <- fit.gam(images)
plot.images(fit, scale = 'raw', masked = FALSE)
plot.images(fit, scale = 'scaled', masked = TRUE)
}
}

