% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.ring.R
\name{plot.ring}
\alias{plot.ring}
\title{plot.ring}
\usage{
\method{plot}{ring}(x, which = c("rate", "time", "intensity", "all"),
  resolution = NA, rate.max = max(x$gam$rings$mid_rate),
  time.max = max(x$gam$data$t), intensity.max = 1, ...)
}
\arguments{
\item{x}{CanaogramGAM to plot}

\item{which}{Which plot type to create (rate, time, intensity, or all plots)}

\item{resolution}{Resolution at which to make the plot (can be higher than the data resolution)}

\item{rate.max}{Rate to plot at maximum brightness}

\item{time.max}{Time to plot at maximum brightness}

\item{intensity.max}{Intensity to plot at maximum brightness}

\item{...}{<not used>}
}
\value{
A ggplot2 object of the created plot(s)
}
\description{
plot.ring
}
\examples{
\dontrun{
images <- read.images('./data/Trial-1/one gree_T00', n = 10)
fit <- fit.gam(images)
plot.ring(fit)
plot.ring(processed, which = 'all')
}
}

