% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.images.R
\name{read.images}
\alias{read.images}
\title{readImages}
\usage{
read.images(root, n = NULL, low = 64, clean = TRUE)
}
\arguments{
\item{root}{A directory/file root use to get all images for this eye}

\item{n}{Restrict to only the first n images}

\item{low}{Resolution for processing}

\item{clean}{Clean up after reading and rescaling images}
}
\value{
A CanalogramImages object containing all of the image data
}
\description{
readImages
}
\examples{
\dontrun{
images <- read.images('./data/Trial-1/one gree_T00', n = 5)
plot(images)
}
}

