% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rescale.images.R
\name{rescale.images}
\alias{rescale.images}
\title{rescale.images}
\usage{
rescale.images(obj, resolution)
}
\arguments{
\item{obj}{CanalogramImages object of images to rescale}

\item{resolution}{New resolution}
}
\value{
CanalogramImages object of rescaled images
}
\description{
rescale.images
}
\examples{
\dontrun{
obj <- rescale.images(images, 16)
}
}

