% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit.gam.R
\name{fit.gam}
\alias{fit.gam}
\title{fit.gam}
\usage{
fit.gam(obj, r_bands = 5, theta_band_size = 10, fit.grid = TRUE,
  fit.bands = TRUE)
}
\arguments{
\item{obj}{CanalogramImages object}

\item{r_bands}{Number of radial bands to fit}

\item{theta_band_size}{Degrees per angular band}

\item{fit.grid}{Whether to fit the models to each grid point}

\item{fit.bands}{Whether to fit a global model to bands}
}
\value{
CanaogramGAM object describing the fit
}
\description{
fit.gam
}
\examples{
\dontrun{
images <- read.images('./data/Trial-1/one gree_T00', n = 5)
plot(images)
fit <- fit.gam(images)
plot(fit)
plot.images(fit)
plot.ring(fit)
}
}

